images: add initramfs image generation
We need to provide a way for devices to mount our rootfs regardless of
the installation medium and partitions.
The currently used root=PARTUUID=<uuid> is only working properly on
devices that support GPT. Although a pseudo uuid is generated for MBR,
we can't control it's value from WIC image generation tools and even if
we could, the generated uuid would differ from the GPT one.
Let's add filesystem UUID and anm initramfs and mount root partitions
with root=UUID=<uuid>

This patch only implements the initramfs generation and adds the proper
filesystem uuid to the WIC image for the TI board.
At the moment we add don't add kernel modules in the initramfs so the image is
2.7mb. This is expected to grow if we need modules in different devices

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>