ftpm: Use upstream packages
Up to now we've been using a special port for fTPM, which was compiling
for Linux.
Microsoft has evolved it's fTPM since then and it can now be compiled
with OE. They also added support for OpenSSL.
Let's take one step at a time, switch to the upstream fTPM now. In the
future we can switch from the expliticitly defined WolfSSL to OpenSSL

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>