StandaloneMm: Add a StandaloneMm binary for secure storage solution
Add a StandaloneMm binary for the secure storage solution. This binary
contains the NOR flash driver needed for storage of the UEFI
variables. The binary currently uses static mappings for the NOR flash
and uart peripherals -- these would have to be subsequently changed to
using a dynamic runtime mechanism for getting the mappings from optee
os.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>