qemu: arm64: Update the conf file to boot from NOR flash0
For enabling the secure storage solution, the platform needs to
enumerate the NOR flash0 device, which is done so by passing it as a
boot time parameter. But enumerating the NOR flash device means that
the firmware binary cannot be passed through the -bios option. Make
the change to remove the -bios option so that the NOR flash can be
enumerated. The firmware images are now put on the NOR flash device.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>