optee-client: do not run tee-supplicant as root
Due to device and filesystem permissions, tee-supplicant and TEE
applications currently have to run as root. This is less than optimal.
Introduce groups 'tee' and 'teeclnt' for tee-supplicant and TEE
applications to use, respectively, as well as the 'tee' user (a member
of 'tee' the group) to run tee-supplicant.
- The new user and groups are created in optee-client_%.bbappend by
inheriting the useradd bbclass
- ledge-minimal.bb is updated to add user 'ledge' to the 'teeclnt'
group for convenience (when OP-TEE is enabled)
- Permissions on device nodes /dev/teepriv0 and /dev/tee0 are set via
udev rules, added in optee-client_%.bbappend
- The systemd startup script for tee-supplicant is updated to run as
tee:tee.
Tested on the qemuarm64-secureboot image.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>