optee-client: add teeclnt group for TEE client applications
Due /dev/tee0 permissions, TEE applications currently have to run as root.
Introduce the teeclnt group so that application do not need full root
priviledges to access the TEE.
- the group is created by optee-client_%.bbappend
- ledge-minimal.bb is updated to add user 'ledge' to the 'teeclnt' group
for convenience (when OP-TEE is enabled)
- Permissions on device /dev/tee0 are set via udev rules, added to
optee-client_%.bbappend

Note that tee-supplicant is still run as root for simplicity and because
it seems there is not much security to gain from running it as a normal
user. It would still need CAP_SYS_RAWIO to access RPMB which is quite a
powerful capability already, and running as non-root introduces some
complexity (notably related to setting permissions on /dev/teepriv0 and
the startup of tee-supplicant, i.e., a dependency between a udev rule and
the tee-supplicant systemd service).

Tested on the qemuarm64-secureboot image.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>