init.cryptfs: do not kill tee-suplicant on exit
Reverts commit 6624b6368ec6 ("init.cryptfs: kill tee-supplicant on exit")
and the fixup commit 069b7cf0a8c1 ("init.cryptfs: fix function definition
syntax"). These changes were invalid because on e.g., RockPi 4, the Linux
kernel (tpm_ftpm_tee module) invokes the fTPM early TA, which in turn
needs to access RPMB secure storage via tee-supplicant. When the
supplicant is not running, the following messages are printed on the
console:

 E/TC:? 0 TA panicked with code 0xffff000c
 E/LD: Status of TA bc50d971-d4c9-42c4-82cb-343fb7f37896
 ...
 [  121.556681] tpm tpm0: ftpm_tee_tpm_op_send: SUBMIT_COMMAND invoke error: 0xffff3024
 ...
 E/TC:? 0 TA panicked with code 0xffff0007
 E/LD: Status of TA bc50d971-d4c9-42c4-82cb-343fb7f37896
 ...
 [  121.806982] optee-ftpm optee-ta-bc50d971-d4c9-42c4-82cb-343fb7f37896: ftpm_tee_probe: tee_client_open_session failed, err=ffff3024
 [  121.808036] optee-ftpm: probe of optee-ta-bc50d971-d4c9-42c4-82cb-343fb7f37896 failed with error -22

The iniramfs tee-supplicant instance needs to be stopped later in the
boot sequence. This problem was not detected on QEMU which doesn't use the
fTPM (yet).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>