optee-client_%.bbappend: unload tpm_ftpm_tee module before stopping tee-supplicant
If the tpm_ftpm_tee module is unloaded from the kernel when tee-supplicant
is not running, the fTPM TA fails to complete some command and there is a
TA panic dump on the console. This happens for example on system shutdown
or reboot. To avoid this, make sure that the module is unloaded before
stopping tee-supplicant. Tested on RockPi 4.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>