init.cryptfs: stop /usr/sbin/rngd on exit
On my RockPi 4 board, 'systemctl halt' or 'systemctl reboot' pauses for
10 seconds when trying to stop /usr/sbin/rngd. This daemon is started
by init.cryptfs in the initramfs and is not even present in the final
root fs, so I suspect its scope should be limited to the initramfs
operations. Kill it before leaving init.cryptfs so that halt and reboot
proceed without delay.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>